(feature) raise if revoking login with no id
Calling this with no user id causes confusing errors since it sends a
delete request to /permissions, which isn't a valid endpoint.

+++ lib/facebook_graph_api/api.rb
@@ -9,6 +9,8 @@ module FacebookGraphApi
     end
 
     def revoke_login(user_id)
+      raise ArgumentError, 'missing user id' if user_id.nil?
+
       http_client.delete("/#{user_id}/permissions")
     end
